The initial commit
Includes the current working version of the terminal emulator. TUI is incomplete and only has some framework classes that serve as a reminder to me in the future. The terminal emulator needs a major architecture overhaul and has many (many) anti-patterns and poor readability. These will be modified in the future. For now, the basics are complete (including a greedy line-wrapping algorithm and text rendering stuff). in the future, I will add character-based color information and better font rendering.