build tools
I've added some build tools to the project; however, I am no Python
programmer, so there are likely many problems that I will need to
address in the future. The resources that is being used in the test is
now included. The filenames in the test have been modified a little to
accomadate the normal directory structure for javac. (I have no idea
what Eclipse was doing in the first place.)