architectural enhancements
So I don't really know what I'm doing here, but hopefully these changes
will make it easier in the future to think about how I'm going to
handle components in the terminal. The TextRenderer interface should be
compatible with BufferedTextPane, but I haven't got around to actally
implementing it. The line in the sand between Swing and the TUI is not
quite clear, and many of these classes may be subject to change once I
work it all out.