compliance with new interface requirements, layout updates....
Yeah, this is a mess right now. I think some things need to be
revisited, such as Layout parameters and what not. I'm not sure I like
how you have to declare interface components now, but the reason why
these changes have been made is so that I can verify that they are
created properly. Before, you needed to be sure to call certain methods
before others, which, in my opinion, leaves too much burden on a client
of this API.